+++ src/ser/mod.rs
@@ -48,17 +48,24 @@ impl fmt::Display for Error {
     }
 }
 
-pub(crate) struct Serializer<'a> {
+/// A structure that serializes Rust values as JSON into a buffer.
+pub struct Serializer<'a> {
     buf: &'a mut [u8],
     current_length: usize,
 }
 
 impl<'a> Serializer<'a> {
-    fn new(buf: &'a mut [u8]) -> Self {
+    /// Create a new `Serializer`
+    pub fn new(buf: &'a mut [u8]) -> Self {
         Serializer {
             buf,
             current_length: 0,
         }
+    }
+
+    /// Return the current amount of serialized data in the buffer
+    pub fn end(&self) -> usize {
+        self.current_length
     }
 
     fn push(&mut self, c: u8) -> Result<()> {
@@ -475,49 +482,14 @@ impl ser::Error for Error {
     }
 }
 
-pub(crate) enum Unreachable {}
-
-impl ser::SerializeTupleStruct for Unreachable {
-    type Ok = ();
-    type Error = Error;
-
-    fn serialize_field<T: ?Sized>(&mut self, _value: &T) -> Result<()> {
-        unreachable!()
-    }
-
-    fn end(self) -> Result<Self::Ok> {
-        unreachable!()
-    }
-}
+/// An unreachable type to fill the SerializeTupleVariant type
+pub enum Unreachable {}
 
 impl ser::SerializeTupleVariant for Unreachable {
     type Ok = ();
     type Error = Error;
 
     fn serialize_field<T: ?Sized>(&mut self, _value: &T) -> Result<()> {
-        unreachable!()
-    }
-
-    fn end(self) -> Result<Self::Ok> {
-        unreachable!()
-    }
-}
-
-impl ser::SerializeMap for Unreachable {
-    type Ok = ();
-    type Error = Error;
-
-    fn serialize_key<T: ?Sized>(&mut self, _key: &T) -> Result<()>
-    where
-        T: ser::Serialize,
-    {
-        unreachable!()
-    }
-
-    fn serialize_value<T: ?Sized>(&mut self, _value: &T) -> Result<()>
-    where
-        T: ser::Serialize,
-    {
         unreachable!()
     }